Enhancement for JRUBY-231, allow implementing JavaBean-like interfaces with attr* methods.
Actually this commit includes logic to allow get/set methods to be implemented with their property names, boolean properties to be implemented with their question-marked versions, and other boolean methods to be implemented with their question-marked versions, all either Java-case or Ruby-case. Whether this full set of features will remain through 1.1.4 is unknown, but it's there for the moment.


git-svn-id: https://svn.codehaus.org/jruby/trunk/jruby@7413 961051c9-f516-0410-bf72-c9f7e237a7b7