Fix for JRUBY-3012: Rake tasks are empty in Mr. Bones' generated project under JRuby
The actual problem here was in repeated reopenings of stdio streams. The logic to dup2 the incoming descriptor's contents "into" the stdio descriptor was wiping out the original stdio fileno. As a result, future reopens did not follow the descriptor-preserving path and the stdio stream ended up getting closed. Because we can't really dup stdio descriptors on JVM, we have separate logic for reopen to prevent this, but it is keyed off fileno and so failed to fire the second time.


git-svn-id: https://svn.codehaus.org/jruby/trunk/jruby@7794 961051c9-f516-0410-bf72-c9f7e237a7b7