Fix for JRUBY-3036: Module include bug with Test::Unit (ObjectSpace issue)
The problem here is that when walking the subclasses, it was not filtering out included pseudo-classes used for mixin inheritance. 
Before, these classes did not show up in the subclasses lists, but we needed to add it for the new serial-number cache invalidation 
logic (which really needs a consistent bidir hierarchy.


git-svn-id: https://svn.codehaus.org/jruby/trunk/jruby@7830 961051c9-f516-0410-bf72-c9f7e237a7b7