Fix for JRUBY-2993: implementing java class causes exception
The problem here was that implementing an interface with multiple 64-bit primitive arguments was corrupting the stack. Specifically, it was always doing "load" instructions at 32-bit boundaries (incrementing the load index by 1 each time). For a single value this appeared to work ok, because it increments the index once and then does no further local variable access. But with two values, the second lload or dload loads from the middle of the first value, resulting in the register error. I fixed it by making sure the index += 2 when the current load is an lload or dload.

git-svn-id: https://svn.codehaus.org/jruby/trunk/jruby@7883 961051c9-f516-0410-bf72-c9f7e237a7b7