Fix for JRUBY-2975: Overriding Time._dump does not behave the same as MRI
The fix for this was reasonably simple. MRI handles load as follows:

* If there are ivars (leading 'I' in marshal output), it sets a flag and proceeds to unmarshal the actual object
* For new-style object unmarshalling (marshal_load and marshal_dump), the flag is ignored; when the object returns, the flag is still set and Marshal fills in the ivars from the stream
* For old-style object unmarshalling (_load and _dump), it is the responsibility of the base _load to initialize the object and then copy all ivars from the "data" string.

In our case, we always did the ivar unmarshalling after the object had been completely unmarshalled, which broke in the Rails example because it expected the "original load" to return an object with ivars intact. Since we didn't install the ivars until after _load had returned, it broke.

The modification was to stuff the ivars into the data string and to modify Time's _load to copy them out, as in the MRI code. There could be other outstanding cases where we are not properly copying variables out of the data string passed to _load, but at least this case is now working correctly.

git-svn-id: https://svn.codehaus.org/jruby/trunk/jruby@7978 961051c9-f516-0410-bf72-c9f7e237a7b7