Improvement for JRUBY-3152: Process.times returns invalid values
This improvement tries to use the built-in management capabilities of the JVM to tally up all running threads. We could probably do a slightly more accurate measurement with the native times function, but this is a good step in the right direction.

If the bean is not available or thread timings are not supported, it falls back on the dumb "ms since started" version.

git-svn-id: https://svn.codehaus.org/jruby/trunk/jruby@8228 961051c9-f516-0410-bf72-c9f7e237a7b7