Fix JRUBY-3407 FFI Stack Level Too Deep on Library.send(:method, args)
This was caused by the DynamicMethod instance method that took the boxed
args, and a block, calling up to the super method with unboxed args and
a block, which called back to the method with boxed args and a block...

It should have been calling the non-block version.  Fixed by moving all
the implementation into a common method, and having all entry points
call it instead.

git-svn-id: https://svn.codehaus.org/jruby/trunk/jruby@9193 961051c9-f516-0410-bf72-c9f7e237a7b7