Fix for JRUBY-3218: Import is sometimes confused by unquoted class name
This issue was largely because we assumed that only lower-case-leading elements of a package were likely to be package name elements. If there were anything other than a lower-case letter, we went straight to loading it as a class.

This all will be improved once we get the jython package cache integrated, but for now I extended the lower-case smarts to anything without a leading upper-case letter.

git-svn-id: https://svn.codehaus.org/jruby/trunk/jruby@9208 961051c9-f516-0410-bf72-c9f7e237a7b7